feat: IFS-4183 adjusted existing documentation
