+++ docs/modules/ROOT/pages/konzept/changelog.adoc
@@ -1,0 +1,6 @@
+[[changelog]]
+= Changelog
+
+*Änderungen IsyFact 4.0.0*
+
+
