+++ docs/modules/ROOT/pages/konzept/docinfo.adoc
