+++ docs/modules/ROOT/pages/nutzungsvorgaben/changelog.adoc
@@ -1,5 +1,4 @@
 [[changelog]]
 = Changelog
 
-*Änderungen IsyFact 4.0.0*
