--- docs/modules/pages/nutzungsvorgaben/changelog.adoc
@@ -1,50 +1,0 @@
-[[changelog]]
-= Changelog
-
-// *Änderungen IsyFact 3.0.0*
-
-// tag::release-3.0.0[]
-
-// end::release-3.0.0[]
-
-// *Änderungen IsyFact 2.4.0*
-
-// tag::release-2.4.0[]
-
-// end::release-2.4.0[]
-
-// *Änderungen IsyFact 2.3.0*
-
-// tag::release-2.3.0[]
-
-// end::release-2.3.0[]
-
-// *Änderungen IsyFact 2.2.0*
-
-// tag::release-2.2.0[]
-
-// end::release-2.2.0[]
-
-// *Änderungen IsyFact 2.1.0*
-
-// tag::release-2.1.0[]
-
-// end::release-2.1.0[]
-
-// *Änderungen IsyFact 2.0.0*
-
-// tag::release-2.0.0[]
-
-// end::release-2.0.0[]
-
-// *Änderungen IsyFact 1.7.0*
-
-// tag::release-1.7.0[]
-
-// end::release-1.7.0[]
-
-// *Änderungen IsyFact 1.6.0*
-
-// tag::release-1.6.0[]
-
-// end::release-1.6.0[]
