--- docs/modules/pages/nutzungsvorgaben/inhalt.adoc
@@ -1,31 +1,0 @@
-= Nutzungsvorgaben Isy Util: Inhalt
-
-// tag::inhalt[]
-[[aufbau-von-isy-util]]
-== Aufbau von isy-util
-
-[[common]]
-=== Common
-
-Das Package `de.bund.bva.isyfact.util.common` enthält allgemeine Klassen, die keiner speziellen Funktionalität zugeordnet sind.
-Zurzeit besteht sie aus folgender Klasse:
-
-* `RecursiveToStringBuilder`: Diese Klasse erzeugt eine Textausgabe für Objekte, die keine geeignete `toString`-Methode implementieren.
-
-[[package-exception]]
-=== Package exception
-
-In `de.bund.bva.isyfact.util.exception` sind Klassen enthalten, die für die Fehlerbehandlung einzusetzen sind:
-
-* `MessageSourceFehlertextProvider`: Durch diese Klasse wird der `FehlertextProvider` implementiert, der sich im Package `de.bund.bva.isyfact.exception` der xref:glossary:glossary:master.adoc#glossar-bibliothek[Bibliothek] isy-exception-core befindet.
-Der `FehlertextProvider` beschreibt Methoden zum Auslesen von Fehlertexten auf Basis von IDs der Ausnahmen und Parameter, sofern welche vorhanden sind.
-
-[[package-spring]]
-=== Package spring
-
-Das Package `de.bund.bva.isyfact.util.spring` enthält Werkzeuge für den Umgang mit Spring:
-
-* `MessageSourceHolder`: Der `MessageSourceHolder` ist eine Referenz auf die aktuelle `MessageSource`-Bean.
-Zusätzlich bietet er Convenience-Funktionen zum Auslesen von `ResourceBundle`-Einträgen aus der `MessageSource`.
-
-// end::inhalt[]
