+++ src/main/java/de/bund/bva/isyfact/util/persistence/package.html
@@ -17,5 +17,5 @@
 
 -->
 <body>
-Enth&auml;lt allgemeine Klassen f&uuml;r die Fehlerbehandlung.
+Utility classes for persistence.
 </body>
