+++ src/main/java/de/bund/bva/isyfact/util/text/FehlerSchluessel.java
@@ -14,7 +14,7 @@
  * implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
-package de.bund.bva.isyfact.util.messages;
+package de.bund.bva.isyfact.util.text;
 
 /**
  * Fehlerschlüssel für isy-util.
