+++ src/main/java/de/bund/bva/isyfact/util/text/package.html
@@ -17,5 +17,5 @@
 
 -->
 <body>
-Enthält Klassen, die keiner speziellen Funktionalität zugeordnet werden können.
+Classes for the creation of texts and messages.
 </body>
