+++ src/test/java/de/bund/bva/isyfact/util/text/MyNode.java
@@ -14,23 +14,11 @@
  * implied. See the License for the specific language governing
  * permissions and limitations under the License.
  */
+package de.bund.bva.isyfact.util.text;
 
-package de.bund.bva.isyfact.util.messages;
+public class MyNode {
 
-/**
- * Diese Klasse implementiert einen Provider für Texte, der zum Laden der Nachrichten den
- * {@link MessageSourceHolder} verwendet. Letztere muss als Spring-Bean in der Anwendung konfiguriert sein.
- */
-public class MessageSourceMessageProvider {
-
-    /**
-     * Returns the message to the given key.
-     *
-     * @param schluessel The key.
-     * @param parameter  Optional parameters.
-     * @return The message to the given key.
-     */
-    public String getMessage(String schluessel, String... parameter) {
-        return MessageSourceHolder.getMessage(schluessel, parameter);
-    }
+	public MyNode child;
+	
+	public String str;
 }
